Refactor class Game to own file
