+++ src/Spaceobject.h
@@ -10,9 +10,9 @@ class Spaceobject: public Size, public Position
     public:
         Spaceobject(size_t, size_t, Sprite&);
     //private:
-    Sprite &obj_sprite;
-    size_t lives;
-    int move_direction;
+        Sprite &obj_sprite;
+        size_t lives;
+        int move_direction;
 };
 
 #endif
