+++ src/main.cpp
@@ -6,38 +6,9 @@
 #include "Size.h"
 #include "Position.h"
 #include "Sprites.h"
-#include "Buffer.h"
 #include "Spaceobject.h"
-
-
-class Game: public Size
-{
-    public:
-        Game(size_t width, size_t height, Sprites &sprites):
-            Size(width, height),
-            player(Spaceobject(width / 2 - 5, 30, sprites.player_sprite))
-        {
-            for (size_t xi = 0; xi < 11; ++xi) {
-                for (size_t yi = 0; yi < 5; ++yi) {
-                    const size_t alien_type = (5-yi) / 2 + 1;
-
-                    Sprite& alien_sprite = sprites.alien_sprites[2 * alien_type - 1];
-                    const size_t xpos = 24 + 16 * xi + (13 - alien_sprite.width)/2;
-                    const size_t ypos = 128 + 17 * yi;
-
-                    aliens.push_back(Spaceobject(xpos, ypos, alien_sprite));
-                }
-            }
-        }
-        ~Game(void)
-        {
-            //
-        }
-
-    //private:
-        Spaceobject player;
-        std::vector<Spaceobject> aliens;
-};
+#include "Buffer.h"
+#include "Game.h"
 
 
 int main(void)
@@ -48,8 +19,6 @@ int main(void)
 
     Game game(buff_width, buff_height, sprites);
 
-    //SpaceObject player;
-    //player = new SpaceObject(100, 100);
 
     uint8_t r = (uint8_t) 0;
     uint8_t g = (uint8_t) 0;
